make method reusable
